Finish making changes to parse 10x hdf5 files and write them to disk.
