+++ code/Makefile
@@ -55,7 +55,7 @@ dmlrun: src/dmlsearch.cpp $(wildcard include/fgc/*.h)
 
 
 HDFLAGS=-L$(HDFPATH)/lib -I$(HDFPATH)/include -lhdf5_cpp -lhdf5 -lhdf5_hl -lhdf5_hl_cpp
-hdf2sparseblz: src/hdf2sparseblz.cpp $(wildcard include/fgc/*.h)
+hdf2dm: src/hdf2dm.cpp $(wildcard include/fgc/*.h)
 	ls $(HDFPATH)/lib && \
 	$(CXX) $(CXXFLAGS) $< -o $@ -DNDEBUG -fopenmp -O3 $(HDFLAGS)
